backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
requests

